# swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Please read on https://nordkern.dk/wp-content/uploads/2018/03/Projektplan.pdf",
    "version" : "0.0.1",
    "title" : "soeofficer",
    "termsOfService" : "Please read on https://nordkern.dk/wp-content/uploads/2018/03/Projektplan.pdf",
    "contact" : {
      "name" : "Nordkern ApS",
      "url" : "https://nordkern.dk",
      "email" : "kontakt@nordkern.dk"
    }
  },
  "basePath" : "/api",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/relations" : {
      "post" : {
        "tags" : [ "relations" ],
        "summary" : "Add a new relation to the system",
        "description" : "",
        "operationId" : "createRelation",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "parent",
          "in" : "query",
          "description" : "The parent of the relation",
          "required" : true,
          "type" : "object"
        }, {
          "name" : "child",
          "in" : "query",
          "description" : "The child of the relation",
          "required" : true,
          "type" : "object"
        }, {
          "name" : "title",
          "in" : "query",
          "description" : "The title of the parent. I.e. {Mother,Father}",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Relation"
            }
          }
        }
      }
    },
    "/relations/{id}" : {
      "put" : {
        "tags" : [ "relations" ],
        "summary" : "Update an existing relation of the system",
        "description" : "",
        "operationId" : "updateRelation",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "parent",
          "in" : "query",
          "description" : "The parent of the relation",
          "required" : false,
          "type" : "object"
        }, {
          "name" : "child",
          "in" : "query",
          "description" : "The child of the relation",
          "required" : false,
          "type" : "object"
        }, {
          "name" : "title",
          "in" : "query",
          "description" : "The title of the parent. I.e. {Mother,Father}",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Relation"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "relations" ],
        "summary" : "Delete relation from the system",
        "description" : "",
        "operationId" : "deleteRelation",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Person" : {
      "type" : "object",
      "required" : [ "dateOfBirth", "gender", "givenName", "id", "surname" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the person"
        },
        "givenName" : {
          "type" : "string",
          "example" : "John",
          "description" : "The given name of the person"
        },
        "surname" : {
          "type" : "string",
          "example" : "Doe",
          "description" : "The surname of the person"
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY"
        },
        "gender" : {
          "type" : "string",
          "example" : "Mand",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "enum" : [ "Male", "Female", "Unknown" ]
        },
        "dateOfDeath" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date of death. Format is: DD/MM/YYYY"
        }
      }
    },
...
    "Relation" : {
      "type" : "object",
      "required" : [ "child", "id", "parent" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the relation"
        },
        "parent" : {
          "example" : "1",
          "description" : "The parent of the relation",
          "$ref" : "#/definitions/Person"
        },
        "child" : {
          "example" : "1",
          "description" : "The parent of the relation",
          "$ref" : "#/definitions/Person"
        }
      }
    },
...
  }
}